requirements.txt
asgiref==3.7.2
channels==4.0.0
contourpy==1.2.0
crispy-bootstrap4==2024.1
cycler==0.12.1
Django==5.0.2
django-bootstrap-calendar==0.1.1
django-braces==1.15.0
django-crispy-forms==2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
Faker==23.3.0
fonttools==4.49.0
install==1.3.5
jsmin==3.0.1
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tzdata==2024.1
